+++ core/src/main/scala/com/evolutiongaming/kafka/flow/timer/TimersOf.scala
@@ -3,6 +3,7 @@ package com.evolutiongaming.kafka.flow.timer
 import cats.effect.Sync
 import cats.syntax.all._
 import com.evolutiongaming.catshelper.Log
+import com.evolutiongaming.catshelper.LogOf
 
 trait TimersOf[F[_], K] {
 
@@ -12,6 +13,7 @@ trait TimersOf[F[_], K] {
 
 object TimersOf {
 
+  @deprecated("use version with `LogOf` to minimze number of required implicits", "0.3.3")
   def memory[F[_]: Sync: Log, K]: TimersOf[F, K] = { (key, createdAt) =>
     Timestamps.of(createdAt) flatMap { implicit timestamps =>
       Timers.memory(key) map { timers =>
@@ -19,5 +21,14 @@ object TimersOf {
       }
     }
   }
+
+  def memory[F[_]: Sync: LogOf, K]: F[TimersOf[F, K]] =
+    LogOf[F].apply(TimersOf.getClass) map { implicit log => (key, createdAt) =>
+      Timestamps.of(createdAt) flatMap { implicit timestamps =>
+        Timers.memory(key) map { timers =>
+          TimerContext(timers, timestamps)
+        }
+      }
+    }
 
 }