+++ docs/changelog.md
@@ -34,6 +34,9 @@ def keyStateWithMetrics = keyStateOf.withCollectorRegistry[IO](???)
 ```
 - `PartitionFlowOf.eagerRecoveryKafkaPersistence` lost `keyStateOfTransform`
 parameter used to construct metrics as the metric is provided out of the box.
+- `KeyDatabase.keysOf`, `SnapshotDatabase.snapshotsOf` and
+`JournalDatabase.journalsOf` require `LogOf` instead of `Log` to minimize number
+of passed implicits.
 
 ## 0.2.x